finish create_final_acteur reorg : ISO
