+++ dags/compute_acteurs/tasks/business_logic/deduplicate_labels.py
@@ -1,0 +1,15 @@
+import pandas as pd
+from compute_acteurs.tasks.transform.merge_and_deduplicate import (
+    deduplicate_acteurs_many2many_relationship,
+)
+
+
+def deduplicate_labels(
+    df_children: pd.DataFrame,
+    df_merge_labels: pd.DataFrame,
+):
+    return deduplicate_acteurs_many2many_relationship(
+        df_children,
+        df_merge_labels,
+        "labelqualite_id",
+    )
